+++ src/router/Helpers.js
@@ -1,4 +1,4 @@
-import types from '../types'
+import types from '../../types'
 // import { firstUpperCase, atLeastOneMatch } from '../Helpers.js'
 import { atLeastOneMatch } from '../Helpers.js'
